requirements.txt
fastapi==0.128.0
SQLAlchemy==2.0.45
uvicorn==0.40.0
alembic==1.17.2
python-jose==3.5.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
ffmpeg-python==0.2.0
sqlmodel==0.0.31
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
libtorrent==2.0.11